Refactor hardware detection utilities
